\name{bujar-package}
\alias{bujar-package}
\docType{package}
\title{
Title: Buckley-James Regression with High-Dimensional Biomarker Data
}
\description{
Buckley-James regression for right-censoring survival data
}
\details{
Buckley-James regression for right-censoring survival data with high-dimensional covariates. Including L_2 boosting with componentwise linear least squares, componentwise smoothing splines, P-splines, regression trees and boosted MARS. Other high-dimensional tools include elastic net, MARS, ACOSSO.
\tabular{ll}{
Package: \tab bujar\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2012-12-20\cr
License: \tab GPL (version 2 or newer)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhu Wang
Maintainer: Zhu Wang <zwang@ccmckids.org>
}
\references{
 Zhu Wang and C.Y. Wang (2010),
 Buckley-James Boosting for Survival Analysis with High-Dimensional
Biomarker Data. 
  \emph{Statistical Applications in Genetics and Molecular Biology},
  Vol. 9 : Iss. 1, Article 24.
}
\keyword{ package }
